requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
black==23.1.0
click==8.1.3
crispy-bootstrap5==0.7
Django==4.1.7
django-bootstrap-datepicker-plus==5.0.3
django-crispy-forms==2.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pydantic==1.10.5
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.5.0
